REF: move misplaced pd.concat tests
